feat(assembleExportDiscoveryMetaData): Moved business logic to separate file to improve maintainability
